requirements.txt
flask
cryptography
pyOpenSSL 